Add files via upload
There are the csv files, along with the script (honor's.py) to make 1000 copies of the csv file. Please run the script first before the ipynb (honor's project KNN algorithm with foreign exchange data set.ipynb). These 1000 copies of the csv files will be become one big file with 20000 records in order for the algorithm to work.